refactor/EV-804/mudando estado para riverpod e componetizando list-coin
