--- lib/my_app.dart
@@ -1,16 +1,0 @@
-import 'package:card_1/pages/moedas_pages.dart';
-import 'package:flutter/material.dart';
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: 'Flutter Demo',
-      theme: ThemeData(),
-      home: const MoedasPageWidgte(),
-    );
-  }
-}
